creo2urdf: divide code in hdr and c++ and move some functions in dedicated sources
