+++ src/creo2urdf/include/creo2urdf/Creo2Urdf.h
@@ -5,3 +5,25 @@
  * This software may be modified and distributed under the terms of the
  * BSD-3-Clause license. See the accompanying LICENSE file for details.
  */
+
+#include <pfcGlobal.h>
+
+class Creo2UrdfActionListerner : public pfcUICommandActionListener {
+public:
+    void OnCommand() override;
+};
+
+class Creo2UrdfAccessListener : public pfcUICommandAccessListener {
+public:
+    pfcCommandAccess OnCommandAccess(xbool AllowErrorMessages) override;
+};
+
+class ValidatorListener : public pfcUICommandActionListener {
+public:
+    void OnCommand() override;
+};
+
+class ValidatorAccessListener : public pfcUICommandAccessListener {
+public:
+    pfcCommandAccess OnCommandAccess(xbool AllowErrorMessages) override;
+};
