+++ src/creo2urdf/src/Creo2Urdf.cpp
@@ -7,619 +7,317 @@
  */
 #define _USE_MATH_DEFINES
 
-#include <pfcGlobal.h>
-#include <pfcModel.h>
-#include <pfcSolid.h>
-#include <pfcShrinkwrap.h>
-#include <pfcAssembly.h>
-#include <pfcComponentFeat.h>
-
-#include <wfcGeometry.h>
-#include <wfcModelItem.h>
-
-#include <cmath>
-#include <string>
-#include <array>
-#include <map>
+#include <creo2urdf/Creo2Urdf.h>
+#include <creo2urdf/Utils.h>
 
 
-#include <iDynTree/Model/Model.h>
-#include <iDynTree/ModelIO/ModelExporter.h>
-#include <iDynTree/ModelIO/ModelLoader.h>
-#include <iDynTree/Model/RevoluteJoint.h>
-#include <iDynTree/KinDynComputations.h>
-#include <iDynTree/Model/Traversal.h>
+void Creo2UrdfActionListerner::OnCommand() {
 
-
-constexpr double mm_to_m   = 1e-3;
-constexpr double mm2_to_m2 = 1e-6;
-constexpr double epsilon   = 1e-12;
-constexpr double rad2deg   = 57.295779513;
-constexpr double deg2rad   = 0.017453293;
-
-
-std::array<std::string, 5> relevant_csys_names = {
-"SCSYS_NECK_1",
-"SCSYS_NECK_2",
-"SCSYS_NECK_3", 
-"SCSYS_HEAD", 
-"SCSYS_REALSENSE"
-};
-
-std::map<std::string, std::string> child_axis_map = {
-    {"SIM_ECUB_HEAD_NECK_2", "NECK_PITCH_AXIS"},
-    {"SIM_ECUB_HEAD_NECK_3", "NECK_ROLL_AXIS"},
-    {"SIM_ECUB_HEAD", "NECK_YAW_AXIS"},
-    {"SIM_ECUB_REALSENSE", "REALSENSE_PITCH_AXIS"}
-};
-
-enum class c2uLogLevel
-{
-    NONE = 0,
-    INFO,
-    WARN,
-    PROMPT
-};
-
-const std::map<c2uLogLevel, std::string> log_level_key = {
-    {c2uLogLevel::NONE, "c2uNONE"},
-    {c2uLogLevel::INFO, "c2uINFO"},
-    {c2uLogLevel::WARN, "c2uWARN"},
-    {c2uLogLevel::PROMPT, "c2uPROMPT"}
-};
-
-std::array<double, 3> computeUnitVectorFromAxis(pfcCurveDescriptor_ptr axis_data)
-{
-    auto axis_line = pfcLineDescriptor::cast(axis_data); // cursed cast from hell
-
-    // There are just two points in the array
-    pfcPoint3D_ptr pstart = axis_line->GetEnd1();
-    pfcPoint3D_ptr pend = axis_line->GetEnd2();
-
-    std::array<double, 3> unit_vector = { 0, 0, 0 };
-
-    double module = sqrt(pow(pend->get(0) - pstart->get(0), 2) +
-        pow(pend->get(1) - pstart->get(1), 2) +
-        pow(pend->get(2) - pstart->get(2), 2));
-
-    if (module < epsilon)
-    {
-        return unit_vector;
-    }
-
-    unit_vector[0] = (pend->get(0) - pstart->get(0)) / module;
-    unit_vector[1] = (pend->get(1) - pstart->get(1)) / module;
-    unit_vector[2] = (pend->get(2) - pstart->get(2)) / module;
-
-    return unit_vector;
-}
-
-iDynTree::SpatialInertia fromCreo(pfcMassProperty_ptr mass_prop, iDynTree::Transform H)
-{
-    auto com = mass_prop->GetGravityCenter();
-    auto inertia_tensor = mass_prop->GetCenterGravityInertiaTensor();
-    iDynTree::RotationalInertiaRaw idyn_inertia_tensor = iDynTree::RotationalInertiaRaw::Zero();
-
-    for (int i_row = 0; i_row < idyn_inertia_tensor.rows(); i_row++) {
-        for (int j_col = 0; j_col < idyn_inertia_tensor.cols(); j_col++) {
-            idyn_inertia_tensor.setVal(i_row, j_col, inertia_tensor->get(i_row, j_col) * mm2_to_m2);
-        }
-    }
-
-    iDynTree::Position com_child({ com->get(0) * mm_to_m , com->get(1) * mm_to_m, com->get(2) * mm_to_m });
-    com_child = H.inverse() * com_child;  // TODO verify
-
-    iDynTree::SpatialInertia sp_inertia(mass_prop->GetMass(), com_child, idyn_inertia_tensor);
-    sp_inertia.fromRotationalInertiaWrtCenterOfMass(mass_prop->GetMass(), com_child, idyn_inertia_tensor);
-
-    return sp_inertia;
-}
-
-iDynTree::Transform fromCreo(pfcTransform3D_ptr creo_trf)
-{
-    iDynTree::Transform idyn_trf;
-    auto o = creo_trf->GetOrigin();
-    auto m = creo_trf->GetMatrix();
-    idyn_trf.setPosition({ o->get(0) * mm_to_m, o->get(1) * mm_to_m, o->get(2) * mm_to_m });
-    idyn_trf.setRotation({ m->get(0,0), m->get(1,0), m->get(2,0),
-                           m->get(0,1), m->get(1,1), m->get(2,1),
-                           m->get(0,2), m->get(1,2), m->get(2,2) });
-
-    return idyn_trf;
-}
-
-void printToMessageWindow(std::string message, c2uLogLevel log_level = c2uLogLevel::INFO)
-{
     pfcSession_ptr session_ptr = pfcGetProESession();
-    xstringsequence_ptr msg_sequence = xstringsequence::create();
-    msg_sequence->append(xstring(message.c_str()));
-    session_ptr->UIClearMessage();
-    session_ptr->UIDisplayMessage("creo2urdf.txt", log_level_key.at(log_level).c_str(), msg_sequence);
-}
+    pfcModel_ptr model_ptr = session_ptr->GetCurrentModel();
+    pfcSolid_ptr solid_ptr = pfcSolid::cast(session_ptr->GetCurrentModel());
 
-void printTransformMatrix(pfcTransform3D_ptr m)
-{
-    printToMessageWindow(fromCreo(m).toString());
-}
+    // TODO Principal units probably to be changed from MM to M before getting the model properties
+    //auto length_unit = solid_ptr->GetPrincipalUnits()->GetUnit(pfcUnitType::pfcUNIT_LENGTH);
+    // length_unit->Modify(pfcUnitConversionFactor::Create(0.001), length_unit->GetReferenceUnit()); // IT DOES NOT WORK
 
-void printRotationMatrix(pfcMatrix3D_ptr m)
-{
-    printToMessageWindow(to_string(m->get(0, 0)) + " " + to_string(m->get(0, 1)) + " " + to_string(m->get(0, 2)));
-    printToMessageWindow(to_string(m->get(1, 0)) + " " + to_string(m->get(1, 1)) + " " + to_string(m->get(1, 2)));
-    printToMessageWindow(to_string(m->get(2, 0)) + " " + to_string(m->get(2, 1)) + " " + to_string(m->get(2, 2)));
-}
+    iDynTree::Model idyn_model;
+    iDynTree::ModelExporterOptions export_options;
+    export_options.robotExportedName = "ECUB_HEAD";
+    export_options.baseLink = "SIM_ECUB_HEAD_NECK_1";
 
-
-
-
-void sanitizeSTL(std::string stl)
-{
-    size_t n_bytes = 5;
-    char placeholder[6] = "robot";
-    std::ofstream output(stl, std::ios::binary | std::ios::out | std::ios::in);
-    
-    for (size_t i = 0; i < n_bytes; i++)
-    {
-        output.seekp(i);
-        output.write(&placeholder[i], 1);
-    }
-    output.close();
-}
-
-std::pair<bool, iDynTree::Transform> getTransformFromPart(pfcModel_ptr modelhdl, const std::string& link_child_name) {
-
-    iDynTree::Transform H_child;
-
-    auto csys_list = modelhdl->ListItems(pfcModelItemType::pfcITEM_COORD_SYS);
-    if (csys_list->getarraysize() == 0) {
-        printToMessageWindow("There are no CSYS in the part " + link_child_name, c2uLogLevel::WARN);
-        
-        H_child = iDynTree::Transform::Identity();
-        
-        return { false, H_child };
+    auto asm_component_list = model_ptr->ListItems(pfcModelItemType::pfcITEM_FEATURE);
+    if (asm_component_list->getarraysize() == 0) {
+        printToMessageWindow("There are no FEATURES in the asm", c2uLogLevel::WARN);
+        return;
     }
 
-    for (size_t i = 0; i < csys_list->getarraysize(); i++)
+    std::ofstream idyn_model_out("iDynTreeModel.txt");
+    std::ofstream idyn_error_out("iDynTreeErrors.txt");
+    std::streambuf* cerr_old_buf = std::cerr.rdbuf(); // store the original cerr buffer
+    std::cerr.rdbuf(idyn_error_out.rdbuf());          // Pass file buffer to set it as new buffer
+
+    std::string link_parent_name = "";
+
+    iDynTree::Transform H_parent = iDynTree::Transform::Identity();
+
+    int component_counter = 0;
+    bool ret;
+
+    for (int i = 0; i < asm_component_list->getarraysize(); i++)
     {
-        auto csys_elem = csys_list->get(i);
+        iDynTree::Link link_child;
+        iDynTree::Transform H_child = iDynTree::Transform::Identity();
+        iDynTree::Transform H_parent_to_child = iDynTree::Transform::Identity();
+        auto comp = asm_component_list->get(i);
+        auto feat = pfcFeature::cast(comp);
+        // auto feat_id = feat->GetId();
 
-        auto csys = pfcCoordSystem::cast(csys_elem);
 
-        if (std::find(relevant_csys_names.begin(), relevant_csys_names.end(), string(csys->GetName())) == relevant_csys_names.end())
+        if (feat->GetFeatType() != pfcFeatureType::pfcFEATTYPE_COMPONENT)
         {
             continue;
         }
 
-        auto trf = csys->GetCoordSys();
+        auto modelhdl = session_ptr->RetrieveModel(pfcComponentFeat::cast(feat)->GetModelDescr());
+        auto link_child_name = string(modelhdl->GetFullName());
 
-        auto m = trf->GetMatrix();
-        auto o = trf->GetOrigin();
+        printToMessageWindow(link_child_name);
 
-        H_child = fromCreo(trf);
-        
-        /*
-        printToMessageWindow("csys name " + string(csys->GetName()));
-        printToMessageWindow("origin x: " + to_string(o->get(0)) + " y: " + to_string(o->get(1)) + " z: " + to_string(o->get(2)));
-        printToMessageWindow("transform:");
-        printToMessageWindow(to_string(m->get(0, 0)) + " " + to_string(m->get(0, 1)) + " " + to_string(m->get(0, 2)));
-        printToMessageWindow(to_string(m->get(1, 0)) + " " + to_string(m->get(1, 1)) + " " + to_string(m->get(1, 2)));
-        printToMessageWindow(to_string(m->get(2, 0)) + " " + to_string(m->get(2, 1)) + " " + to_string(m->get(2, 2)));
-        */
-        //printToMessageWindow(string(csys_feat->GetFeatTypeName()));
+        xintsequence_ptr seq = xintsequence::create();
+        seq->append(feat->GetId());
 
-        return { true, H_child };
-    }
+        pfcComponentPath_ptr comp_path = pfcCreateComponentPath(pfcAssembly::cast(model_ptr), seq);
 
-    return { false, H_child };
-}
+        auto asm_csys_H_csys = fromCreo(comp_path->GetTransform(xtrue));
 
-std::pair<bool, iDynTree::Direction> getRotationAxisFromPart(pfcModel_ptr modelhdl, const std::string& link_child_name, iDynTree::Transform H_child) {
+        iDynTree::Transform csys_H_child;
+        std::tie(ret, csys_H_child) = getTransformFromPart(modelhdl, link_child_name);
 
-    iDynTree::Direction axis_unit_vector;
+        H_child = asm_csys_H_csys * csys_H_child;
 
-    auto axes_list = modelhdl->ListItems(pfcModelItemType::pfcITEM_AXIS);
-    // printToMessageWindow("There are " + to_string(axes_list->getarraysize()) + " axes");
-    if (axes_list->getarraysize() == 0) {
-        printToMessageWindow("There is no AXIS in the part " + link_child_name, c2uLogLevel::WARN);
-
-        axis_unit_vector.zero();
-        return { false, axis_unit_vector};
-    }
-
-    pfcAxis* axis = nullptr;
-
-    for (size_t i = 0; i < axes_list->getarraysize(); i++)
-    {
-        auto axis_elem = pfcAxis::cast(axes_list->get(i));
-
-        if (string(axis_elem->GetName()) == child_axis_map.at(string(link_child_name)))
+        if (!ret)
         {
-            axis = axis_elem;
-            // printToMessageWindow("The axis is called " + string(axis_elem->GetName()));
-        }
-    }
-
-    auto axis_data = wfcWAxis::cast(axis)->GetAxisData();
-
-    auto axis_line = pfcLineDescriptor::cast(axis_data); // cursed cast from hell
-
-    auto unit = computeUnitVectorFromAxis(axis_line);
-
-    axis_unit_vector.setVal(0, unit[0]);
-    axis_unit_vector.setVal(1, unit[1]);
-    axis_unit_vector.setVal(2, unit[2]);
-
-    axis_unit_vector = H_child.inverse() * axis_unit_vector;  // We might benefit from performing this operation directly in Creo
-    axis_unit_vector.Normalize();
-
-    /*
-    printToMessageWindow(string(axis->GetName()) + ": (" + std::to_string(axis_unit_vector[0]) + ", "
-                                                        + std::to_string(axis_unit_vector[1]) + ", "
-                                                        + std::to_string(axis_unit_vector[2]) + ")");
-    */
-    return { true, axis_unit_vector };
-}
-
-bool addMeshAndExport(pfcModel_ptr modelhdl, const std::string& link_child_name, int component_counter, iDynTree::Model& idyn_model)
-{
-    //printToMessageWindow("Using " + relevant_csys_names[component_counter] + " to make stl");
-
-    std::string stl_file_name = link_child_name + ".stl";
-
-    // Make all alphabetic characters lowercase
-    std::transform(stl_file_name.begin(), stl_file_name.end(), stl_file_name.begin(),
-                    [](unsigned char c) { return std::tolower(c); });
-
-    modelhdl->Export(stl_file_name.c_str(), pfcExportInstructions::cast(pfcSTLBinaryExportInstructions().Create(relevant_csys_names[component_counter].c_str())));
-
-    // Replace the first 5 bytes of the binary file with a string different than "solid"
-    // to avoid issues with stl parsers.
-    // For details see: https://github.com/icub-tech-iit/creo2urdf/issues/16
-    sanitizeSTL(string(link_child_name) + ".stl");
-
-    // Lets add the mesh to the link
-    iDynTree::ExternalMesh visualMesh;
-    // Meshes are in millimeters, while iDynTree models are in meters
-    iDynTree::Vector3 scale; scale(0) = scale(1) = scale(2) = mm_to_m;
-    visualMesh.setScale(scale);
-    // Let's assign a gray as default color
-    iDynTree::Vector4 color;
-    iDynTree::Material material;
-    color(0) = color(1) = color(2) = 0.5;
-    color(3) = 1.0;
-    material.setColor(color);
-    visualMesh.setMaterial(material);
-    // Assign transform
-    // TODO Right now maybe it is not needed it ie exported respct the link csys
-    // visualMesh.setLink_H_geometry(H_parent_to_child);
-
-    // Assign name
-    visualMesh.setFilename(stl_file_name);
-    // TODO Right now let's consider visual and collision with the same mesh
-    idyn_model.visualSolidShapes().getLinkSolidShapes()[idyn_model.getLinkIndex(string(link_child_name))].push_back(visualMesh.clone());
-    idyn_model.collisionSolidShapes().getLinkSolidShapes()[idyn_model.getLinkIndex(string(link_child_name))].push_back(visualMesh.clone());
-
-    return true;
-}
-
-class Creo2UrdfActionListerner : public pfcUICommandActionListener {
-public:
-    void OnCommand() override {
-
-        pfcSession_ptr session_ptr = pfcGetProESession();
-        pfcModel_ptr model_ptr = session_ptr->GetCurrentModel();
-        pfcSolid_ptr solid_ptr = pfcSolid::cast(session_ptr->GetCurrentModel());
-
-        // TODO Principal units probably to be changed from MM to M before getting the model properties
-        //auto length_unit = solid_ptr->GetPrincipalUnits()->GetUnit(pfcUnitType::pfcUNIT_LENGTH);
-        // length_unit->Modify(pfcUnitConversionFactor::Create(0.001), length_unit->GetReferenceUnit()); // IT DOES NOT WORK
-
-        iDynTree::Model idyn_model;
-        iDynTree::ModelExporterOptions export_options;
-        export_options.robotExportedName = "ECUB_HEAD";
-        export_options.baseLink = "SIM_ECUB_HEAD_NECK_1";
-
-        auto asm_component_list = model_ptr->ListItems(pfcModelItemType::pfcITEM_FEATURE);
-        if (asm_component_list->getarraysize() == 0) {
-            printToMessageWindow("There are no FEATURES in the asm", c2uLogLevel::WARN);
+            std::cerr.rdbuf(cerr_old_buf);
             return;
         }
 
-        std::ofstream idyn_model_out("iDynTreeModel.txt");
-        std::ofstream idyn_error_out("iDynTreeErrors.txt");
-        std::streambuf* cerr_old_buf = std::cerr.rdbuf(); // store the original cerr buffer
-        std::cerr.rdbuf(idyn_error_out.rdbuf());          // Pass file buffer to set it as new buffer
+        auto mass_prop = pfcSolid::cast(modelhdl)->GetMassProperty();
+        auto com = mass_prop->GetGravityCenter();                     // TODO transform the center of mass in relative coords
+        auto comInertia = mass_prop->GetCenterGravityInertiaTensor(); // TODO GetCoordSysInertia ?
 
-        std::string link_parent_name = "";
+        link_child.setInertia(fromCreo(mass_prop, H_child));
 
-        iDynTree::Transform H_parent = iDynTree::Transform::Identity();
-
-        int component_counter = 0;
-        bool ret;
-
-        for (int i = 0; i < asm_component_list->getarraysize(); i++)
+        if (!link_child.getInertia().isPhysicallyConsistent())
         {
-            iDynTree::Link link_child;
-            iDynTree::Transform H_child = iDynTree::Transform::Identity();
-            iDynTree::Transform H_parent_to_child = iDynTree::Transform::Identity();
-            auto comp = asm_component_list->get(i);
-            auto feat = pfcFeature::cast(comp);
-            // auto feat_id = feat->GetId();
+            printToMessageWindow(link_child_name + " is NOT physically consistent!", c2uLogLevel::WARN);
+        }
 
+        /*
+        printToMessageWindow("Model name is " + std::string(name) + " and weighs " + to_string(mass_prop->GetMass()));
+        printToMessageWindow("Center of mass: x: " + to_string(com->get(0)) + " y: " + to_string(com->get(1)) + " z: " + to_string(com->get(2)));
+        printToMessageWindow("Inertia tensor:");
+        printToMessageWindow(to_string(comInertia->get(0, 0)) + " " + to_string(comInertia->get(0, 1)) + " " + to_string(comInertia->get(0, 2)));
+        printToMessageWindow(to_string(comInertia->get(1, 0)) + " " + to_string(comInertia->get(1, 1)) + " " + to_string(comInertia->get(1, 2)));
+        printToMessageWindow(to_string(comInertia->get(2, 0)) + " " + to_string(comInertia->get(2, 1)) + " " + to_string(comInertia->get(2, 2)));
+        */
 
-            if (feat->GetFeatType() != pfcFeatureType::pfcFEATTYPE_COMPONENT)
-            {
-                continue;
-            }
-
-            auto modelhdl = session_ptr->RetrieveModel(pfcComponentFeat::cast(feat)->GetModelDescr());
-            auto link_child_name = string(modelhdl->GetFullName());
-
-            printToMessageWindow(link_child_name);
-
-            xintsequence_ptr seq = xintsequence::create();
-            seq->append(feat->GetId());
-
-            pfcComponentPath_ptr comp_path = pfcCreateComponentPath(pfcAssembly::cast(model_ptr), seq);
-
-            auto asm_csys_H_csys = fromCreo(comp_path->GetTransform(xtrue));
-
-            iDynTree::Transform csys_H_child;
-            std::tie(ret, csys_H_child) = getTransformFromPart(modelhdl, link_child_name);
-
-            H_child = asm_csys_H_csys * csys_H_child;
+        if (component_counter > 0)
+        {
+            iDynTree::Direction axis;
+            std::tie(ret, axis) = getRotationAxisFromPart(modelhdl, link_child_name, H_child);
 
             if (!ret)
             {
                 std::cerr.rdbuf(cerr_old_buf);
+                return; 
+            }
+
+            H_parent_to_child = H_parent.inverse() * H_child;
+
+            //printToMessageWindow("H_parent: " + H_parent.toString());
+            //printToMessageWindow("H_child: " + H_child.toString());
+            //printToMessageWindow("prev_link_H_link: " + H_parent_to_child.toString());
+
+            iDynTree::RevoluteJoint joint(H_parent_to_child, {axis, H_parent_to_child.getPosition()});
+                                                                // Should be 0 the origin of the axis, the displacement is already considered in transform
+                                                                //{ o->get(0) * mm_to_m, o->get(1) * mm_to_m, o->get(2) * mm_to_m } });
+
+            // TODO let's put the limits hardcoded, to be retrieved from Creo
+            double min = 0.0;
+            double max = M_PI;
+            joint.enablePosLimits(true);
+            joint.setPosLimits(0, min, max);
+            // TODO we have to retrieve the rest transform from creo
+            //joint.setRestTransform();
+
+            if (idyn_model.addJointAndLink(link_parent_name, link_parent_name + "--" +link_child_name, &joint, link_child_name, link_child) == iDynTree::JOINT_INVALID_INDEX) {
+                printToMessageWindow("FAILED TO ADD JOINT!", c2uLogLevel::WARN);
+
+                std::cerr.rdbuf(cerr_old_buf);
                 return;
             }
-
-            auto mass_prop = pfcSolid::cast(modelhdl)->GetMassProperty();
-            auto com = mass_prop->GetGravityCenter();                     // TODO transform the center of mass in relative coords
-            auto comInertia = mass_prop->GetCenterGravityInertiaTensor(); // TODO GetCoordSysInertia ?
-
-            link_child.setInertia(fromCreo(mass_prop, H_child));
-
-            if (!link_child.getInertia().isPhysicallyConsistent())
-            {
-                printToMessageWindow(link_child_name + " is NOT physically consistent!", c2uLogLevel::WARN);
-            }
-
-            /*
-            printToMessageWindow("Model name is " + std::string(name) + " and weighs " + to_string(mass_prop->GetMass()));
-            printToMessageWindow("Center of mass: x: " + to_string(com->get(0)) + " y: " + to_string(com->get(1)) + " z: " + to_string(com->get(2)));
-            printToMessageWindow("Inertia tensor:");
-            printToMessageWindow(to_string(comInertia->get(0, 0)) + " " + to_string(comInertia->get(0, 1)) + " " + to_string(comInertia->get(0, 2)));
-            printToMessageWindow(to_string(comInertia->get(1, 0)) + " " + to_string(comInertia->get(1, 1)) + " " + to_string(comInertia->get(1, 2)));
-            printToMessageWindow(to_string(comInertia->get(2, 0)) + " " + to_string(comInertia->get(2, 1)) + " " + to_string(comInertia->get(2, 2)));
-            */
-
-            if (component_counter > 0)
-            {
-                iDynTree::Direction axis;
-                std::tie(ret, axis) = getRotationAxisFromPart(modelhdl, link_child_name, H_child);
-
-                if (!ret)
-                {
-                    std::cerr.rdbuf(cerr_old_buf);
-                    return; 
-                }
-
-                H_parent_to_child = H_parent.inverse() * H_child;
-
-                //printToMessageWindow("H_parent: " + H_parent.toString());
-                //printToMessageWindow("H_child: " + H_child.toString());
-                //printToMessageWindow("prev_link_H_link: " + H_parent_to_child.toString());
-
-                iDynTree::RevoluteJoint joint(H_parent_to_child, {axis, H_parent_to_child.getPosition()});
-                                                                   // Should be 0 the origin of the axis, the displacement is already considered in transform
-                                                                   //{ o->get(0) * mm_to_m, o->get(1) * mm_to_m, o->get(2) * mm_to_m } });
-
-                // TODO let's put the limits hardcoded, to be retrieved from Creo
-                double min = 0.0;
-                double max = M_PI;
-                joint.enablePosLimits(true);
-                joint.setPosLimits(0, min, max);
-                // TODO we have to retrieve the rest transform from creo
-                //joint.setRestTransform();
-
-                if (idyn_model.addJointAndLink(link_parent_name, link_parent_name + "--" +link_child_name, &joint, link_child_name, link_child) == iDynTree::JOINT_INVALID_INDEX) {
-                    printToMessageWindow("FAILED TO ADD JOINT!", c2uLogLevel::WARN);
-
-                    std::cerr.rdbuf(cerr_old_buf);
-                    return;
-                }
-                printToMessageWindow(to_string(component_counter) + ": " + link_parent_name + "--" + link_child_name);
-            }
-            else
-            {
-                printToMessageWindow("First link, skipping joint addition");
-                idyn_model.addLink(string(link_child_name), link_child);
-            }
-
-            addMeshAndExport(modelhdl, link_child_name, component_counter, idyn_model);
-
-            link_parent_name = link_child_name;
-            H_parent = H_child;
-            component_counter++;
-
-            // TODO when we have an additional frame to add
-            // idyn_model.addAdditionalFrameToLink(string(name), string(name) + "_" + string(csys_list->get(0)->GetName()), fromCreo(transform)); 
-        }
-
-        idyn_model_out << idyn_model.toString();
-        idyn_model_out.close();
-        iDynTree::ModelExporter mdl_exporter;
-        mdl_exporter.init(idyn_model);
-        mdl_exporter.setExportingOptions(export_options);
-
-        if (!mdl_exporter.isValid())
-        {
-            printToMessageWindow("Model is not valid!", c2uLogLevel::WARN);
-        }
-
-        /*
-        if (idyn_model.getLinkIndex("SIM_ECUB_HEAD_NECK_1") == iDynTree::LINK_INVALID_INDEX)
-        {
-            printToMessageWindow("[ERROR] URDFStringFromModel: specified baseLink is not part of the model");
-        }
-
-       iDynTree::Traversal exportTraversal;
-        if (!idyn_model.computeFullTreeTraversal(exportTraversal, idyn_model.getLinkIndex("SIM_ECUB_HEAD_NECK_1")))
-        {
-            printToMessageWindow("[ERROR] URDFStringFromModel: error in computeFullTreeTraversal", c2uLogLevel::WARN);
+            printToMessageWindow(to_string(component_counter) + ": " + link_parent_name + "--" + link_child_name);
         }
         else
         {
-            printToMessageWindow("Traversal completed successfully");
-        }
-        */
-
-        if (!mdl_exporter.exportModelToFile("model.urdf"))
-        {
-            printToMessageWindow("Error exporting the urdf. See iDynTreeErrors.txt for details", c2uLogLevel::WARN);
-        }
-        else
-        {
-            printToMessageWindow("Urdf created successfully!");
+            printToMessageWindow("First link, skipping joint addition");
+            idyn_model.addLink(string(link_child_name), link_child);
         }
 
-        std::cerr.rdbuf(cerr_old_buf); // Restore original cerr buffer after using iDynTree
-        return;
+        addMeshAndExport(modelhdl, link_child_name, component_counter, idyn_model);
+
+        link_parent_name = link_child_name;
+        H_parent = H_child;
+        component_counter++;
+
+        // TODO when we have an additional frame to add
+        // idyn_model.addAdditionalFrameToLink(string(name), string(name) + "_" + string(csys_list->get(0)->GetName()), fromCreo(transform)); 
     }
-};
+
+    idyn_model_out << idyn_model.toString();
+    idyn_model_out.close();
+    iDynTree::ModelExporter mdl_exporter;
+    mdl_exporter.init(idyn_model);
+    mdl_exporter.setExportingOptions(export_options);
+
+    if (!mdl_exporter.isValid())
+    {
+        printToMessageWindow("Model is not valid!", c2uLogLevel::WARN);
+    }
+
+    /*
+    if (idyn_model.getLinkIndex("SIM_ECUB_HEAD_NECK_1") == iDynTree::LINK_INVALID_INDEX)
+    {
+        printToMessageWindow("[ERROR] URDFStringFromModel: specified baseLink is not part of the model");
+    }
+
+    iDynTree::Traversal exportTraversal;
+    if (!idyn_model.computeFullTreeTraversal(exportTraversal, idyn_model.getLinkIndex("SIM_ECUB_HEAD_NECK_1")))
+    {
+        printToMessageWindow("[ERROR] URDFStringFromModel: error in computeFullTreeTraversal", c2uLogLevel::WARN);
+    }
+    else
+    {
+        printToMessageWindow("Traversal completed successfully");
+    }
+    */
+
+    if (!mdl_exporter.exportModelToFile("model.urdf"))
+    {
+        printToMessageWindow("Error exporting the urdf. See iDynTreeErrors.txt for details", c2uLogLevel::WARN);
+    }
+    else
+    {
+        printToMessageWindow("Urdf created successfully!");
+    }
+
+    std::cerr.rdbuf(cerr_old_buf); // Restore original cerr buffer after using iDynTree
+    return;
+}
 
 
-class Creo2UrdfAccessListener : public pfcUICommandAccessListener {
-public:
-    pfcCommandAccess OnCommandAccess(xbool AllowErrorMessages) override {
-        auto model = pfcGetProESession()->GetCurrentModel();
-        if (!model) {
-            return pfcCommandAccess::pfcACCESS_AVAILABLE;
-        }
-        auto type = model->GetType();
-        if (type != pfcMDL_PART && type != pfcMDL_ASSEMBLY) {
-            return pfcCommandAccess::pfcACCESS_UNAVAILABLE;
-        }
+
+pfcCommandAccess Creo2UrdfAccessListener::OnCommandAccess(xbool AllowErrorMessages) {
+    auto model = pfcGetProESession()->GetCurrentModel();
+    if (!model) {
         return pfcCommandAccess::pfcACCESS_AVAILABLE;
     }
-};
+    auto type = model->GetType();
+    if (type != pfcMDL_PART && type != pfcMDL_ASSEMBLY) {
+        return pfcCommandAccess::pfcACCESS_UNAVAILABLE;
+    }
+    return pfcCommandAccess::pfcACCESS_AVAILABLE;
+}
 
+void ValidatorListener::OnCommand() {
+    std::map<std::string, iDynTree::Transform> link_name_to_creo_computed_trf_map;
+    iDynTree::ModelLoader mdl_loader;
+    if (!mdl_loader.loadModelFromFile("model.urdf"))
+    {
+        printToMessageWindow("Could not load urdf for validation!", c2uLogLevel::WARN);
+        return;
+    }
 
-class ValidatorListener : public pfcUICommandActionListener {
-public:
-    void OnCommand() override {
-        std::map<std::string, iDynTree::Transform> link_name_to_creo_computed_trf_map;
-        iDynTree::ModelLoader mdl_loader;
-        if (!mdl_loader.loadModelFromFile("model.urdf"))
+    auto idyn_model = mdl_loader.model();
+    auto creo_session_ptr = pfcGetProESession();
+    auto creo_model_ptr = creo_session_ptr->GetCurrentModel();
+    bool ret{ false };
+    iDynTree::VectorDynSize positions(idyn_model.getNrOfDOFs());
+    for (int i = 0; i < idyn_model.getNrOfJoints(); i++) {
+        printToMessageWindow("Please insert value for joint " + idyn_model.getJointName(i), c2uLogLevel::PROMPT);
+        positions[i] = creo_session_ptr->UIReadRealMessage(-360.0, 360.0) * deg2rad;
+        printToMessageWindow("joint" + to_string(i) + " value " + to_string(positions[i]), c2uLogLevel::INFO);
+    }
+
+    iDynTree::Transform H_child = iDynTree::Transform::Identity();
+
+    auto asm_component_list = creo_model_ptr->ListItems(pfcModelItemType::pfcITEM_FEATURE);
+    if (asm_component_list->getarraysize() == 0) {
+        printToMessageWindow("There are no FEATURES in the asm", c2uLogLevel::WARN);
+        return;
+    }
+    // Let's first retrieve the trfs from root to all the links and store it in a map.
+    for (int i = 0; i < asm_component_list->getarraysize(); i++)
+    {
+
+        auto feat = pfcFeature::cast(asm_component_list->get(i));
+
+        if (feat->GetFeatType() != pfcFeatureType::pfcFEATTYPE_COMPONENT)
         {
-            printToMessageWindow("Could not load urdf for validation!", c2uLogLevel::WARN);
+            continue;
+        }
+
+        xintsequence_ptr seq = xintsequence::create();
+        seq->append(feat->GetId());
+
+        pfcComponentPath_ptr comp_path = pfcCreateComponentPath(pfcAssembly::cast(creo_model_ptr), seq);
+
+        auto asm_csys_H_csys = fromCreo(comp_path->GetTransform(xtrue));
+        auto modelhdl = creo_session_ptr->RetrieveModel(pfcComponentFeat::cast(feat)->GetModelDescr());
+        auto link_child_name = string(modelhdl->GetFullName());
+
+        iDynTree::Transform csys_H_child;
+        std::tie(ret, csys_H_child) = getTransformFromPart(modelhdl, link_child_name);
+        if (!ret)
+        {
+            printToMessageWindow("Unable to get the transform respect to the root for" + link_child_name);
             return;
         }
 
-        auto idyn_model = mdl_loader.model();
-        auto creo_session_ptr = pfcGetProESession();
-        auto creo_model_ptr = creo_session_ptr->GetCurrentModel();
-        bool ret{ false };
-        iDynTree::VectorDynSize positions(idyn_model.getNrOfDOFs());
-        for (int i = 0; i < idyn_model.getNrOfJoints(); i++) {
-            printToMessageWindow("Please insert value for joint " + idyn_model.getJointName(i), c2uLogLevel::PROMPT);
-            positions[i] = creo_session_ptr->UIReadRealMessage(-360.0, 360.0) * deg2rad;
-            printToMessageWindow("joint" + to_string(i) + " value " + to_string(positions[i]), c2uLogLevel::INFO);
-        }
+        H_child = iDynTree::Transform::compose(asm_csys_H_csys,csys_H_child);
 
-        iDynTree::Transform H_child = iDynTree::Transform::Identity();
+        link_name_to_creo_computed_trf_map[link_child_name] = H_child;
 
-        auto asm_component_list = creo_model_ptr->ListItems(pfcModelItemType::pfcITEM_FEATURE);
-        if (asm_component_list->getarraysize() == 0) {
-            printToMessageWindow("There are no FEATURES in the asm", c2uLogLevel::WARN);
-            return;
-        }
-        // Let's first retrieve the trfs from root to all the links and store it in a map.
-        for (int i = 0; i < asm_component_list->getarraysize(); i++)
-        {
-
-            auto feat = pfcFeature::cast(asm_component_list->get(i));
-
-            if (feat->GetFeatType() != pfcFeatureType::pfcFEATTYPE_COMPONENT)
-            {
-                continue;
-            }
-
-            xintsequence_ptr seq = xintsequence::create();
-            seq->append(feat->GetId());
-
-            pfcComponentPath_ptr comp_path = pfcCreateComponentPath(pfcAssembly::cast(creo_model_ptr), seq);
-
-            auto asm_csys_H_csys = fromCreo(comp_path->GetTransform(xtrue));
-            auto modelhdl = creo_session_ptr->RetrieveModel(pfcComponentFeat::cast(feat)->GetModelDescr());
-            auto link_child_name = string(modelhdl->GetFullName());
-
-            iDynTree::Transform csys_H_child;
-            std::tie(ret, csys_H_child) = getTransformFromPart(modelhdl, link_child_name);
-            if (!ret)
-            {
-                printToMessageWindow("Unable to get the transform respect to the root for" + link_child_name);
-                return;
-            }
-
-            H_child = iDynTree::Transform::compose(asm_csys_H_csys,csys_H_child);
-
-            link_name_to_creo_computed_trf_map[link_child_name] = H_child;
-
-        }
-        printToMessageWindow("Running model validation");
-        iDynTree::KinDynComputations computer;
-        computer.loadRobotModel(idyn_model);
-
-
-        // Setting the `world_T_base`
-        iDynTree::Vector3 gravity; gravity.zero(); gravity(2) = -9.8;
-        auto H_base = link_name_to_creo_computed_trf_map.at("SIM_ECUB_HEAD_NECK_1");
-
-
-        if (!computer.setRobotState(H_base, positions,
-             iDynTree::Twist(),
-             iDynTree::VectorDynSize(idyn_model.getNrOfDOFs()),
-             gravity))
-        {
-            printToMessageWindow("Could not set the robot state!", c2uLogLevel::WARN);
-        }
-
-        // Lets ask the same transform to idyntree and check the difference w/ the
-        // creo computed one
-        for (int i = 0; i < idyn_model.getNrOfLinks(); i++) {
-            auto link_name = idyn_model.getLinkName(i);
-            auto idyn_trf = computer.getWorldTransform(link_name);
-            auto creo_trf = link_name_to_creo_computed_trf_map.at(link_name);
-
-            printToMessageWindow("Position error for "+ link_name + ": "+ (idyn_trf.getPosition() - creo_trf.getPosition()).toString());
-            printToMessageWindow("RPY error for " + link_name + ": " +
-                                std::to_string(idyn_trf.getRotation().asRPY()[0] - creo_trf.getRotation().asRPY()[0]) + ", " +
-                                std::to_string(idyn_trf.getRotation().asRPY()[1] - creo_trf.getRotation().asRPY()[1]) + ", " +
-                                std::to_string(idyn_trf.getRotation().asRPY()[2] - creo_trf.getRotation().asRPY()[2]));
-            //printToMessageWindow("IDYN TRANSFORM: " + idyn_trf.toString());
-            //printToMessageWindow("CREO TRANSFORM: " + creo_trf.toString());
-        }
-        return;
     }
-};
+    printToMessageWindow("Running model validation");
+    iDynTree::KinDynComputations computer;
+    computer.loadRobotModel(idyn_model);
 
-class ValidatorAccessListener : public pfcUICommandAccessListener {
-public:
-    pfcCommandAccess OnCommandAccess(xbool AllowErrorMessages) override {
-        auto model = pfcGetProESession()->GetCurrentModel();
-        if (!model) {
-            return pfcCommandAccess::pfcACCESS_AVAILABLE;
-        }
-        auto type = model->GetType();
-        if (type != pfcMDL_PART && type != pfcMDL_ASSEMBLY) {
-            return pfcCommandAccess::pfcACCESS_UNAVAILABLE;
-        }
+
+    // Setting the `world_T_base`
+    iDynTree::Vector3 gravity; gravity.zero(); gravity(2) = -9.8;
+    auto H_base = link_name_to_creo_computed_trf_map.at("SIM_ECUB_HEAD_NECK_1");
+
+
+    if (!computer.setRobotState(H_base, positions,
+            iDynTree::Twist(),
+            iDynTree::VectorDynSize(idyn_model.getNrOfDOFs()),
+            gravity))
+    {
+        printToMessageWindow("Could not set the robot state!", c2uLogLevel::WARN);
+    }
+
+    // Lets ask the same transform to idyntree and check the difference w/ the
+    // creo computed one
+    for (int i = 0; i < idyn_model.getNrOfLinks(); i++) {
+        auto link_name = idyn_model.getLinkName(i);
+        auto idyn_trf = computer.getWorldTransform(link_name);
+        auto creo_trf = link_name_to_creo_computed_trf_map.at(link_name);
+
+        printToMessageWindow("Position error for "+ link_name + ": "+ (idyn_trf.getPosition() - creo_trf.getPosition()).toString());
+        printToMessageWindow("RPY error for " + link_name + ": " +
+                            std::to_string(idyn_trf.getRotation().asRPY()[0] - creo_trf.getRotation().asRPY()[0]) + ", " +
+                            std::to_string(idyn_trf.getRotation().asRPY()[1] - creo_trf.getRotation().asRPY()[1]) + ", " +
+                            std::to_string(idyn_trf.getRotation().asRPY()[2] - creo_trf.getRotation().asRPY()[2]));
+        //printToMessageWindow("IDYN TRANSFORM: " + idyn_trf.toString());
+        //printToMessageWindow("CREO TRANSFORM: " + creo_trf.toString());
+    }
+    return;
+}
+
+pfcCommandAccess ValidatorAccessListener::OnCommandAccess(xbool AllowErrorMessages) {
+    auto model = pfcGetProESession()->GetCurrentModel();
+    if (!model) {
         return pfcCommandAccess::pfcACCESS_AVAILABLE;
     }
-};
+    auto type = model->GetType();
+    if (type != pfcMDL_PART && type != pfcMDL_ASSEMBLY) {
+        return pfcCommandAccess::pfcACCESS_UNAVAILABLE;
+    }
+    return pfcCommandAccess::pfcACCESS_AVAILABLE;
+}
+
 
 /*====================================================================*\
 FUNCTION : user_initialize()
